content/smil/test/test_smilCSSPaced.xhtml

<html xmlns="http://www.w3.org/1999/xhtml">
<head>
  <title>Test for Animation Behavior on CSS Properties</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="smilTestUtils.js"></script>
  <script type="text/javascript" src="db_smilCSSPropertyList.js"></script>
  <script type="text/javascript" src="db_smilCSSPaced.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<p id="display"></p>
<div id="content">
<svg xmlns="http://www.w3.org/2000/svg"
     width="200px" height="200px" font-size="50px" style="color: rgb(50,50,50)"
     onload="this.pauseAnimations()">
  <rect x="20" y="20" width="200" height="200"/>
  <text x="20" y="20">testing 123</text>
  <marker/>
</svg>
</div>
<pre id="test">
<script class="testbody" type="text/javascript">
<![CDATA[

SimpleTest.waitForExplicitFinish();

function main()
{
  if (!SMILUtil.isSMILEnabled()) {
    ok(false, "SMIL dosn't seem to be enabled");
    SimpleTest.finish();
    return;
  }

  // Start out with document paused
  var svg = SMILUtil.getSVGRoot();
  ok(svg.animationsPaused(), "should be paused by <svg> load handler");
  is(svg.getCurrentTime(), 0, "should be paused at 0 in <svg> load handler");

  testBundleList(gPacedBundles, new SMILTimingData(1.0, 6.0));
  // Set "display:none" on everything and run the tests again
  SMILUtil.hideSubtree(SMILUtil.getSVGRoot(), false, false);
  testBundleList(gPacedBundles, new SMILTimingData(1.0, 6.0));

  SimpleTest.finish();
}

window.addEventListener("load", main, false);
]]>
</script>
</pre>
</body>
</html>
